Following discussion: explicit key names, prints
- Printing a message if no rules are applied when compile, obey-fix or
obey-check called
- obeyRules key become obeyRulesDir, as it refers to the directory for the
.class
- obeyWarn and obeyFix are now obeyWarnTags and obeyFixTags, as they do not
list the rules, but the tag specified in the project definition.

Currently working on trying to find a way to avoid having to call NSC when
obey-list is called, but it seems that this would require to scan the
obeyRulesDir and process the warn/fix options, i.e. do the exact same job
as what the compiler plugin is doing, but duplicated.