Adding empty "Merge" class with first comments
Adding "Merge" class, which is empty. This will be use to merge trees into
token streams using the original token stream, the original tree and the
modified tree.
The "Print" class could be moved outside of Obey to scalameta at some point.

For development purpose, everyting will be kept there for now.