Removing Keeper, not required as no local rules
Since all rules are imported, the keeper is not required anymore. We simply
removed it. its logic (as well as Holder, renamed as TagHolder) have been
moved to the UserOptions inside the NSC plugin.

Those things remain as vars, as they need to hold the current state based on
the user input.