Adding modifications to messages
Modifications will be useful to reprint the tree properly. Note that
this will be come a MUST-DO: all rewriting rules should also specifiy
the modified tree. This will avoid the overhead to go through the trees
once more to find all modfications.
There is a small issue at runtime with rules defined outside of the scope
of the main Obey project (in Obey-example), which will have to be addressed.