Adding modified tree to rule mess. for transforms
Adding both the modified tree and the original tree to message form
transforming rules, which will allow to properly identify where the
modiciation have happen.

TODO: we will have to check in more details how the rewriting should be done;
it might happen that we will reconstruct the tree without the need of the fully
modified one by TQL finally.