Adding comments on rules, removing ListToSet
Nobody will ever use the ListToSet rule. Expanding the rule to cover
a more general case such as creating a list somewhere and transforming
it into a set afterward is perfectly correct, so there is no reason
to extends it. We simply remove it.

Comment were added in some rules that need to be made more complicated.